Renamed stuff to work with wasm (requires lib and not main)
