+++ Cargo.toml
@@ -8,6 +8,7 @@ edition = "2021"
 [dependencies]
 rand = "0.8.4"
 wasm-bindgen = "0.2.63"
+getrandom = { version = "0.2", features = ["js"] } #required to be explicit for wasm-bindgen to compile.
 
 [lib]
-crate-type = ["cdylib", "rlib"]
+crate-type = ["cdylib"]
